Actualizando test

+++ src/test/java/cucumberJava/borrarAmigoTest.java
@@ -31,7 +31,7 @@ public class borrarAmigoTest {
 	public void Eliminar_un_usuario_que_es_tu_amigo() {
 		borrador=new Usuario("borrador.borrador");
 		borrado1=new Usuario("borrado1.borrado1", "1234qwer", "borrado1@alu.uclm.es", "borrado1");
-		usuarioDao.delete(borrado1);
+		if(usuarioDao.selectNombre(borrado1)) usuarioDao.delete(borrado1);
 		try {		
 			usuarioDao.insert(borrado1);
 			utilidades.enviarSolicitud(borrador, borrado1);
@@ -57,7 +57,7 @@ public class borrarAmigoTest {
 	public void Eliminar_un_usuario_que_no_es_tu_amigo() {
 		borrador=new Usuario("borrador.borrador");
 		borrado2=new Usuario("borrado2.borrado2", "1234qwer", "borrado2@alu.uclm.es", "borrado2");
-		usuarioDao.delete(borrado2);
+		if(usuarioDao.selectNombre(borrado2))usuarioDao.delete(borrado2);
 		try {		
 			usuarioDao.insert(borrado2);
 		} catch (Exception e) {
