+++ src/test/java/cucumberJava/solicitudAceptarTest.java
@@ -19,7 +19,7 @@ public class solicitudAceptarTest {
 	@Given("^Usuario conectado para aceptar$")
 	public void Usuario_conectado_para_aceptar() {
 		solicitante=new Usuario("solicitante.solicitante", "1234qwer", "solicitante@alu.uclm.es", "solicitante");
-		usuarioDao.delete(solicitante);
+		if(usuarioDao.selectNombre(solicitante))usuarioDao.delete(solicitante);
 		try {		
 			usuarioDao.insert(solicitante);
 		} catch (Exception e) {
@@ -29,7 +29,7 @@ public class solicitudAceptarTest {
 	@When("^Acepta solicitud de alguien que le ha mandado$")
 	public void Acepta_solicitud_de_alguien_que_le_ha_mandado() {
 		acepta=new Usuario("acepta.acepta", "1234qwer", "acepta@alu.uclm.es", "acepta");
-		usuarioDao.delete(acepta);
+		if(usuarioDao.selectNombre(acepta))usuarioDao.delete(acepta);
 		try {
 			usuarioDao.insert(acepta);
 		} catch (Exception e) {
@@ -53,7 +53,7 @@ public class solicitudAceptarTest {
 	@When("^Acepta solicitud de alguien que no le ha mandado$")
 	public void Acepta_solicitud_de_alguien_que_no_le_ha_mandado() {
 		acepta2=new Usuario("acepta2.acepta2", "1234qwer", "acepta2@alu.uclm.es", "acepta2");
-		usuarioDao.delete(acepta2);
+		if(usuarioDao.selectNombre(acepta2))usuarioDao.delete(acepta2);
 		try {
 			usuarioDao.insert(acepta2);
 		} catch (Exception e) {
